services/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "regulus-api"
version = "0.1.0"
description = "Regulus FastAPI service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "pydantic-settings>=2.3.4",
  "redis>=5.0.8",
  "rq>=1.16.2",
  "tree-sitter>=0.22.0",
  "tree-sitter-languages>=1.10.2",
  "pgvector>=0.3.2",
  "openai>=1.51.0",
  "sentence-transformers>=3.0.1"
]

[tool.hatch.build.targets.wheel]
packages = ["src/regulus_api"]
